Homework 1: 2.6: Stochastic gradient descent
I had to jump through a couple of hoops to get variable \eta
working, so I wonder whether I might have made a mistake.
Nevertheless variant gradient descent seems to work,
although it feels as if it's harder to make it converge
to a good optimum, and it seems more finicky than batch gradient descent.