Add homework 2 section 2
Homework 2 section 1 concerns observing the existing code, and section 3 will
only be done after watching some more videos.

I'm unsure whether this is what I was supposed to do, given that it
looked like the original `ridge_regression.py` already seemed to
contain the solutions. Of course it might be that this was supposed to
be a warp up to section 3.